Enhance delete network ports of monitoring

+++ inc/networkport.class.php
@@ -150,12 +150,14 @@ class PluginMonitoringNetworkport extends CommonDBTM {
    static function deleteNetworkPort($parm) {
       global $DB;
       
-      $query = "SELECT * FROM `glpi_plugin_monitoring_networkports`
-         WHERE `networkports_id`='".$parm->fields['id']."'";
-      $result = $DB->query($query);
-      while ($data=$DB->fetch_array($result)) {
-         $pmNetworkPort = new PluginMonitoringNetworkport();
-         $pmNetworkPort->delete($data);
+      if ($parm->fields['itemtype'] == 'NetworkEquipment') {
+         $query = "SELECT * FROM `glpi_plugin_monitoring_networkports`
+            WHERE `networkports_id`='".$parm->fields['id']."'";
+         $result = $DB->query($query);
+         while ($data=$DB->fetch_array($result)) {
+            $pmNetworkPort = new PluginMonitoringNetworkport();
+            $pmNetworkPort->delete($data);
+         }
       }
    }
 }
